refactor: redux 관련 파일 분리 (#24)
