+++ app/(base-layout)/layout.tsx
@@ -1,4 +1,4 @@
-import { ReduxProvider } from '@/app/providers/redux-provider'
+import { ReduxProvider } from '@/app/redux/redux-provider'
 import { BaseLayout } from '@/shared/ui/base-layout'
 
 const Layout = ({ children }: { children: React.ReactNode }) => {
