+++ src/app/redux/redux-provider.tsx
@@ -3,7 +3,7 @@
 import { Provider } from 'react-redux'
 import { PersistGate } from 'redux-persist/integration/react'
 
-import { persistor, store } from '../redux/store'
+import { persistor, store } from './store'
 
 const ReduxProvider = ({ children }: { children: React.ReactNode }) => {
   return (
